+++ crates/radicle-cli/src/commands/sync.rs
@@ -345,7 +345,7 @@ fn sync_status(
     const SYMBOL_STATE_UNKNOWN: &str = "•";
 
     let mut table = Table::<5, term::Label>::new(TableOptions::bordered());
-    let mut seeds: Vec<_> = node.seeds(rid)?.into();
+    let mut seeds: Vec<_> = node.seeds_for(rid, [*profile.did()])?.into();
     let local_nid = node.nid()?;
     let aliases = profile.aliases();
 
@@ -522,7 +522,7 @@ pub fn fetch(
         None => {
             // We push nodes that are in our seed list in attempt to fulfill the
             // replicas, if needed.
-            let seeds = node.seeds(rid)?;
+            let seeds = node.seeds_for(rid, [*profile.did()])?;
             let (connected, disconnected) = seeds.partition();
             let candidates = connected
                 .into_iter()