+++ crates/radicle-cli/src/node.rs
@@ -149,7 +149,7 @@ where
 
     let config = match sync::PrivateNetwork::private_repo(&doc) {
         None => {
-            let (synced, unsynced) = node.seeds(rid)?.iter().fold(
+            let (synced, unsynced) = node.seeds_for(rid, [*me])?.iter().fold(
                 (BTreeSet::new(), BTreeSet::new()),
                 |(mut synced, mut unsynced), seed| {
                     if seed.is_synced() {