+++ crates/radicle-cli/tests/commands.rs
@@ -1704,7 +1704,7 @@ fn test_clone_without_seeds() {
     let working = environment.tempdir().join("working");
     let rid = alice.project("heartwood", "Radicle Heartwood Protocol & Stack");
     let mut alice = alice.spawn();
-    let seeds = alice.handle.seeds(rid).unwrap();
+    let seeds = alice.handle.seeds_for(rid, [alice.id]).unwrap();
     let connected = seeds.connected().collect::<Vec<_>>();
 
     assert!(connected.is_empty());